Makes Fixes for the Luxor Demo 8
Fixes the number of test days so there are enough to calculate the SMA
(which caused the failure of the demo). Adds more information for the
user in the output from walk.forward function, so 'phases' of the
process are more clearly visible in the process (without which
walk.forward() output is impossible to read). Moves dev.off() to where
pdf() is produced (earlier in the code of the demo) so that the
chart.Posn() is not immediately closed after being drawn. Substitutes
OS-filesystem-hostile characters (space and colon) in file names that
walkforward is trying to save (which was another point of failure).
Adjusts graph. parameter ask=FALSE, to save precious precious time while
debugging.