Fixes Demo 8 Further (Temp. Hack: ES() to Fix )
User functions must handle all the errors internally ! Cannot handle
data frame column names elsewhere, so this is the best solution -- to
handle exceptions internally. This commit also serves as a backup of
intermediary analysis. PortfReturns() as it turned out, does not produce
errors on short time series, so error handling will be disabled in the
next commit.